rust compiler: support list/map types and improve range loops

+++ tests/machine/x/rust/for_loop.rs
@@ -1,5 +1,5 @@
 fn main() {
-    for i in 1..4 {
+    for i in (1 as i32)..(4 as i32) {
         println!("{:?}", i);
     }
 }
